Fixed all backward calculation memory leak problems

+++ src/stack.c
@@ -3,7 +3,7 @@
 #include "tensor.h"
 
 // Function to create a stack of given capacity.
-Stack *createStack(unsigned long capacity)
+Stack *createStack(unsigned long long capacity)
 {
     Stack *stack = (Stack *)malloc(sizeof(Stack));
     stack->len = 0;
